Move getCurrentUser to auth_utils
