+++ src/utils/ProtectedRoute.jsx
@@ -3,25 +3,13 @@ import { Navigate } from 'react-router-dom';
 import { PropTypes, instanceOf } from 'prop-types';
 
 import { withCookies, cookieKeys, Cookies } from './cookie_utils';
-import { auth } from './auth_utils';
+import { auth, getCurrentUser } from './auth_utils';
 
 // TODO: Make calls to backend to verify user access token
 // const verifyToken = () => {
 //   return true;
 // };
 
-/**
- * Returns the current user synchronously
- * @param {Auth} authInstance
- * @returns The current user (or undefined)
- */
-const getCurrentUser = authInstance =>
-  new Promise((resolve, reject) => {
-    const unsubscribe = authInstance.onAuthStateChanged(user => {
-      unsubscribe();
-      resolve(user);
-    }, reject);
-  });
 /**
  * Protects a route from unauthenticated users
  * @param {Component} children The component the user is trying to access
@@ -33,7 +21,6 @@ const getCurrentUser = authInstance =>
 const ProtectedRoute = ({ children, redirectPath, roles, cookies }) => {
   const currentUser = getCurrentUser(auth);
   if (currentUser && roles.includes(cookies.get(cookieKeys.ROLE))) {
-    // console.log(roles.includes(cookies.get(cookieKeys.ROLE)));
     return children;
   }
   return <Navigate to={redirectPath} />;
